[Fix] Various fixes - fix boards not getting loaded correctly - fix heads on boards having an offset - fix blockface not getting loaded correctly - fix bungee not being able to load messages - fix get cmd exception - fix head place message missing placeholders - fix get message missing a space - fix admin permission not working on proxy - fix gamemode creative not preventing player from finding heads - fix head placement out of board bounds, if head number is greater than board size - fix rest data in board config file by adding TRUNCATE_EXISTING standard option and thus using default options. - fix iterating over existing board blocks wrongly (maxY check) - fix reset of server.properties every startup - fix doesn't update textures even though we are not working with a state snapshot - fix wrong length calculation
Added prefix to proxy message

Added automatic port configuration when the server starts for the first time.
This still isn't a hands-off thing, like you still have to configure online mode / player forwarding on the proxy. This is because we cant just add a part of the velocity config.

Check data version in BoardManager
Highlight the board after creating with particles
Inflated bounding box by 1 to include the whole block. This automatically allows better length calculation and better overlapping checks